+++ packages/spotlight/src/spotlight.js
@@ -191,9 +191,9 @@ const Spotlight = (function () {
 	// An extension point for updating pointer mode based on the current platform.
 	// Currently only webOS
 	function setPlatformPointerMode () {
-		const palmSystem = window.PalmSystem;
-		if (palmSystem && palmSystem.cursor) {
-			setPointerMode(palmSystem.cursor.visibility);
+		const webOSSystem = window.webOSSystem ?? window.PalmSystem;
+		if (webOSSystem && webOSSystem.cursor) {
+			setPointerMode(webOSSystem.cursor.visibility);
 		}
 	}
 
@@ -202,7 +202,8 @@ const Spotlight = (function () {
 			return false;
 		}
 
-		if ((getPointerMode() && !fromPointer) && (getInputType() === 'touch' || (typeof window !== 'undefined' && (!window.PalmSystem || window.PalmSystem.cursor?.visibility)))) {
+		const webOSSystem = window.webOSSystem ?? window.PalmSystem;
+		if ((getPointerMode() && !fromPointer) && (getInputType() === 'touch' || (typeof window !== 'undefined' && (!webOSSystem || webOSSystem.cursor?.visibility)))) {
 			setContainerLastFocusedElement(elem, containerIds);
 			return false;
 		}
@@ -567,11 +568,11 @@ const Spotlight = (function () {
 				window.addEventListener('mouseover', onMouseOver);
 				window.addEventListener('mousemove', onMouseMove);
 
-				if (platform.touch) {
+				if (platform.touchEvent) {
 					window.addEventListener('touchend', onTouchEnd);
 				}
 
-				if (platform.webos) {
+				if (platform.type === 'webos') {
 					window.top.document.addEventListener('webOSMouse', handleWebOSMouseEvent);
 					window.top.document.addEventListener('keyboardStateChange', handleKeyboardStateChangeEvent);
 				}
@@ -605,11 +606,11 @@ const Spotlight = (function () {
 			window.removeEventListener('mouseover', onMouseOver);
 			window.removeEventListener('mousemove', onMouseMove);
 
-			if (platform.touch) {
+			if (platform.touchEvent) {
 				window.removeEventListener('touchend', onTouchEnd);
 			}
 
-			if (platform.webos) {
+			if (platform.type === 'webos') {
 				window.top.document.removeEventListener('webOSMouse', handleWebOSMouseEvent);
 				window.top.document.removeEventListener('keyboardStateChange', handleKeyboardStateChangeEvent);
 			}