+++ packages/webos/CHANGELOG.md
@@ -2,6 +2,58 @@
 
 The following is a curated list of changes in the Enact webos module, newest changes on the top.
 
+## [unreleased]
+
+### Deprecated
+
+- `webos/platform` member `watch` and `legacy`, to be removed in 5.0.0
+
+### Added
+
+- `webos/platform` member `webos` to detect if a webOS platform
+- `webos/platform` member `version` for the detected webOS version
+- `webos/platform` member `chrome` for the detected chrome version
+
+## [4.8.0] - 2024-02-08
+
+No significant changes.
+
+## [4.7.9] - 2023-12-08
+
+No significant changes.
+
+## [4.5.6] - 2023-11-30
+
+No significant changes.
+
+## [4.0.13] - 2022-11-29
+
+No significant changes.
+
+## [4.7.8] - 2023-11-17
+
+No significant changes.
+
+## [4.7.7] - 2023-11-09
+
+No significant changes.
+
+## [4.7.6] - 2023-09-20
+
+No significant changes.
+
+## [4.7.5] - 2023-09-12
+
+No significant changes.
+
+## [4.7.4] - 2023-08-31
+
+No significant changes.
+
+## [4.7.3] - 2023-08-10
+
+No significant changes.
+
 ## [4.7.2] - 2023-07-14
 
 No significant changes.